% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{exportPDF,Rsocialdata-method}
\alias{exportPDF}
\alias{exportPDF,RegLog-method}
\alias{exportPDF,Rsocialdata-method}
\title{Produce a summary of an object in a PDF file}
\usage{
\S4method{exportPDF}{Rsocialdata}(object, pdfSavingName, dateformat,
  page.orientation = "landscape", latexPackages = NULL, keepTex = F,
  openPDF = T, description.chlength = 300, valids.chlength = 40,
  valids.cut.percent = 0.5, sorting = "decreasing", width.id = 0.5,
  width.varname = 1.5, width.description = 10.5, width.n = 0.8,
  width.na = 1.2, width.valids = 5, width.valids.nao.inc = 5,
  width.min = 1.25, width.max = 1.25, width.mean = 1.25,
  width.stddev = 1.25)

exportPDF(object, pdfSavingName, dateformat, page.orientation = "portrait",
  latexPackages = NULL, keepTex = F, openPDF = T, ...)

\S4method{exportPDF}{RegLog}(object, pdfSavingName, graphics = FALSE,
  page.orientation = "portrait", latexPackages = NULL, keepTex = FALSE,
  openPDF = TRUE, ...)
}
\arguments{
\item{object}{the object for which you to write the summary in PDF.}

\item{pdfSavingName}{a \code{character}}

\item{dateformat}{a \code{character}}

\item{description.chlength}{a \code{numeric}}

\item{graphics}{a \code{logical}}

\item{keepTex}{a \code{logical} keep the source of the tex file?}

\item{latexPackages}{a \code{character}. You can add LaTeX packages you need for the compilation here, especially for supporting other languages than english.}

\item{page.orientation}{the orientation of the page, either "portrait" or "landscape".}

\item{sorting}{a \code{character}. If "decreasing" valid cases are sorted in decrease order of their frequencies. If "increase" an increasing order is used.}

\item{valids.chlength}{a \code{numeric}}

\item{valids.cut.percent}{a \code{numeric}. The minimal frequency need for a valid case to be reported in the document.}

\item{width.id}{a \code{numeric}. Width for the variable row index column, in cm.}

\item{width.varname}{a \code{numeric}. Width for variable name column, in cm.}

\item{width.description}{a \code{numeric}. Width for description column, in cm.}

\item{width.n}{a \code{numeric}. Width for number of valid cases column, in cm.}

\item{width.na}{a \code{numeric}. Width for percents of missing values column, in cm.}

\item{width.valids}{a \code{numeric}. For categorical variables, width for the valid cases listing column, in cm.}

\item{width.valids.nao.inc}{a \code{numeric}. For nominal and ordinal variables, allow to increase the width of the valid cases listing column by reducing the description column, in cm.}

\item{width.min}{a \code{numeric}. For quantitative variables, the with of the minimal values column, in cm.}

\item{width.max}{a \code{numeric}. For quantitative variables, the with of the maximal values column, in cm.}

\item{width.mean}{a \code{numeric}. For quantitative variables, the with of the means column, in cm.}

\item{width.stddev}{a \code{numeric}. For quantitative variables, the with of the standard deviations column, in cm.}

\item{openPDF}{a \code{logical} open the PDF after generation completed?}

\item{...}{Other options to pass to the function. For \code{\link{Statdf}} objects the option \code{merge}, with argument 'no', 'left' or 'right' allow to merge or not the p-values columns, either with the next left or right column.}
}
\description{
This generic method intends to produce a summary of the object in a PDF file. You can also keep the tex source to reuse it in your own documents.
}
\section{Methods (by class)}{
\itemize{
\item \code{Rsocialdata}: method for \code{Rsocialdata} objects. Generate a codebook of the database.

\item \code{RegLog}: method for \code{Reglog} objects. Generate a summary of a logistic regression model.
}}

