% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.spss.file}
\alias{get.spss.file}
\title{Import a SPSS database into a \code{Rsocialdata} database.}
\usage{
get.spss.file(file, datadir, variables = NULL, tsvar = character(0),
  ovar = character(0), wvar = character(0),
  use.coding.order = "increasing", max.value.labels = Inf,
  savingName = NULL, lowernames = TRUE, name = NULL,
  description = character(0), exportPDF = TRUE, reencode = "latin1")
}
\arguments{
\item{file}{the full name of the file you want to import}

\item{datadir}{the directory where your SPSS file is. If missing the working directory is used.}

\item{variables}{the list of variables you want to import. If missing all variables in the SPSS file are imported.}

\item{tsvar}{a vector containing the names of variables having to be converted in the \code{TimestampVariable} type. The SPSS timestamp origin "1582-10-14" is used to set the origin of the time variable.}

\item{ovara}{vector containing the names of variables having to be converted in the \code{OrdinalVariable} type.}

\item{wvara}{vector containing the names of variables having to be converted in the \code{WeightingVariable} type.}

\item{use.coding.order}{a character, either \code{"increasing"}, \code{"decreasing"}, \code{"default"}. If \code{"increasing"} the coding order is used to sort the order in the values of the Variable. If \code{"decreasing"} the order is in a decreasing way. If \code{"default"} no sorting is performed. Default is \code{"increasing"}.}

\item{max.value.labels}{depreciated. Will be soon removed.}

\item{savingName}{a character. By setting this argument you produce an export in .Rdata of the SPSS file with the name given.}

\item{lowernames}{a logical. If \code{TRUE} all names are set in lower case.}

\item{name}{a character. The name of the data base. If \code{NULL} the name of the file is used.}

\item{description}{a character. A description of the database.}

\item{exportPDF}{a logical. If \code{TRUE} the summary in PDF of the data base is created.}

\item{reencode}{a logical: should character strings be re-encoded to the current locale. Default is \code{TRUE}. Alternatively character, specifying an encoding to assume.}
}
\value{
a \code{\link{Rsocialdata-class}} object.
}
\description{
The \code{get.spss.file} is the standard function to import a SPSS database (in a \code{.sav} file) into a \code{Rsocialdata} database. The function is based on the \code{\link[foreign]{read.spss}} function. The \code{\link[foreign]{read.spss}} function is run two times to get both coding and labels of both variables and values. The measure of each imported variable (categorical or scale) is automatically detected. If a variable has all its values (valid cases and missing values) labelised, the variable will be set as a categorical variable. If a variable has at least one of its values (valid cases or missing values) with no label, the variable will be set as a scale variable.
}

