% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weighting<-}
\alias{weighting<-}
\title{Set the non-response weights of a \code{Rsocialdata} database}
\usage{
weighting(object) <- value
}
\arguments{
\item{object}{the \code{Rsocialdata} object for which we want to set the map.}

\item{value}{a character, the name of the variable to use as non-response weights.}
}
\description{
This generic method intends to set or replace the non-response weights of a \code{Rsocialdata} database. Weights has to be defined a \code{\link{WeightingVariable}} object.
}

