% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{weighting}
\alias{weighting}
\alias{weighting,Rsocialdata-method}
\alias{weighting<-,Rsocialdata,character-method}
\title{Get the non-response weights of a \code{Rsocialdata} database}
\usage{
weighting(object)

\S4method{weighting}{Rsocialdata}(object)

\S4method{weighting}{Rsocialdata,character}(object) <- value
}
\arguments{
\item{object}{the \code{Rsocialdata} object for which we want to get the map.}
}
\description{
This generic method intends to extract the non-response weights of a \code{Rsocialdata} database.
}
\section{Methods (by class)}{
\itemize{
\item \code{Rsocialdata}: method for \code{Rsocialdata} objects.

\item \code{object = Rsocialdata,value = character}: method for \code{Rsocialdata} objects.
}}

